requirements.txt
alembic==0.9.1
appdirs==1.4.3
click==6.7
enum-compat==0.0.2
eventlet==0.20.1
Flask==0.12
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SocketIO==2.8.6
Flask-SQLAlchemy==2.2
greenlet==0.4.12
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
pymongo==3.4.0
pyparsing==2.2.0
python-editor==1.0.3
python-engineio==1.3.1
python-socketio==1.7.3
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.6
Werkzeug==0.12.1
qiniu==7.1.2
pillow==4.1.1
